feat: secure deletion

+++ src/app/home/home.page.ts
@@ -24,6 +24,7 @@ import { QuizService } from '../services/quiz.service';
 import { Quiz } from '../models/quiz';
 import { AddQuizModalComponent } from './add-quiz.modal';
 import { AuthService } from '../services/auth';
+import { firstValueFrom } from 'rxjs/internal/firstValueFrom';
 
 @Component({
   selector: 'app-home',
@@ -85,11 +86,36 @@ export class HomePage implements OnInit {
   }
 
   async deleteQuiz(id: string) {
-    this.quizService.deleteQuiz(id);
-    this.quizService.getAll().subscribe((quizzes) => {
-      this.quizzes = quizzes;
-    });
+  const user = this.authService.isConnected();
+  if (!user) {
+    alert("Vous devez être connecté pour supprimer un quiz.");
+    return;
   }
+
+  try {
+    const quiz = await firstValueFrom(this.quizService.get(id));
+    if (!quiz) {
+      alert("Quiz introuvable.");
+      return;
+    }
+
+    const isAuthor = quiz.authorId === user.uid;
+    const isAdmin = user.email === 'admin@example.com';
+    if (!isAuthor && !isAdmin) {
+      alert("Vous n'êtes pas autorisé à supprimer ce quiz.");
+      return;
+    }
+
+    await this.quizService.deleteQuiz(id);
+    alert("Quiz supprimé !");
+
+    this.quizzes = await firstValueFrom(this.quizService.getAll());
+
+  } catch (err) {
+    console.error(err);
+    alert("Erreur lors de la suppression du quiz.");
+  }
+}
 
   isLogged() {
     return this.authService.isConnected();
